synthetic-image-generator/requirements.txt
coretex
Pillow
numpy
